More menu enabled and as a separate file. Added larevel configs for JS, make close button an html button for accessibility
